Make clang analyzer happy in regard to 3rd party frameworks and fix another couple of C++ 11-related things.

+++ src/vkhelper.mm
@@ -766,7 +766,7 @@ VKRequest *VKHelper::ProcessRequest(QVariantMap request)
         if (request.contains("parameters")) {
             QVariantMap parameters = request["parameters"].toMap();
 
-            vk_parameters = [NSMutableDictionary dictionaryWithCapacity:parameters.count()];
+            vk_parameters = [NSMutableDictionary dictionaryWithCapacity:static_cast<NSUInteger>(parameters.count())];
 
             foreach (QString key, parameters.keys()) {
                 vk_parameters[key.toNSString()] = parameters[key].toString().toNSString();
